Refactor findChar to utils
- Refactor and move finChar to utils, instead of slashcommands function
- Refactor scrapers to use actual init functionality
